+++ source/CHANGELOG.md
@@ -4,7 +4,7 @@
 Version 2.3.2
 -------------
 NEW:
-* Mouse wheel in chart now zooms
+* Mouse wheel in charts now zooms
 
 CHANGED:
 * Analysis function GoodnessOfFit: more warnings in the log if not all data is usable